Middleware de permissoes

+++ app/Console/Commands/CrudGeneratorCommand.php
@@ -34,11 +34,16 @@ class CrudGeneratorCommand extends Command
         //'App\Http\Controllers\{$name}Controller::class'
         $nameController = $name . "Controller";
         //File::append(base_path('routes/api.php'), "\n Route::resource('".Str::plural(strtolower($name))."'".str_replace(".","",",App\Http\Controllers\.$nameController.::class);"));
-        File::append(base_path('routes/api.php'), "\n \n Route::get('" . Str::plural(strtolower($name)) . "'" . str_replace(".", "", ",[App\Http\Controllers\.$nameController.::class,'index'])->middleware('auth:sanctum');"));
-        File::append(base_path('routes/api.php'), "\n Route::get('" . Str::plural(strtolower($name)) . "/{id}'" . str_replace(".", "", ",[App\Http\Controllers\.$nameController.::class,'show'])->middleware('auth:sanctum');"));
-        File::append(base_path('routes/api.php'), "\n Route::post('" . Str::plural(strtolower($name)) . "'" . str_replace(".", "", ",[App\Http\Controllers\.$nameController.::class,'store'])->middleware('auth:sanctum');"));
-        File::append(base_path('routes/api.php'), "\n Route::put('" . Str::plural(strtolower($name)) . "/{id}'" . str_replace(".", "", ",[App\Http\Controllers\.$nameController.::class,'update'])->middleware('auth:sanctum');"));
-        File::append(base_path('routes/api.php'), "\n Route::delete('" . Str::plural(strtolower($name)) . "/{id}'" . str_replace(".", "", ",[App\Http\Controllers\.$nameController.::class,'delete'])->middleware('auth:sanctum');"));
+        //middleware(['auth:sanctum','access_control'])->name('casas.get');
+        $routeNameGet = Str::plural(strtolower($name))."-get";
+        $routeNamePost = Str::plural(strtolower($name))."-post";
+        $routeNamePut = Str::plural(strtolower($name))."-put";
+        $routeNameDelete = Str::plural(strtolower($name))."-delete";
+        File::append(base_path('routes/api.php'), "\n \n Route::get('" . Str::plural(strtolower($name)) . "'" . str_replace(".", "", ",[App\Http\Controllers\.$nameController.::class,'index'])->middleware(['auth:sanctum','access_control'])->name('".$routeNameGet."');"));
+        File::append(base_path('routes/api.php'), "\n Route::get('" . Str::plural(strtolower($name)) . "/{id}'" . str_replace(".", "", ",[App\Http\Controllers\.$nameController.::class,'show'])->middleware(['auth:sanctum','access_control'])->name('".$routeNameGet."');"));
+        File::append(base_path('routes/api.php'), "\n Route::post('" . Str::plural(strtolower($name)) . "'" . str_replace(".", "", ",[App\Http\Controllers\.$nameController.::class,'store'])->middleware(['auth:sanctum','access_control'])->name('".$routeNamePost."');"));
+        File::append(base_path('routes/api.php'), "\n Route::put('" . Str::plural(strtolower($name)) . "/{id}'" . str_replace(".", "", ",[App\Http\Controllers\.$nameController.::class,'update'])->middleware(['auth:sanctum','access_control'])->name('".$routeNamePut."');"));
+        File::append(base_path('routes/api.php'), "\n Route::delete('" . Str::plural(strtolower($name)) . "/{id}'" . str_replace(".", "", ",[App\Http\Controllers\.$nameController.::class,'delete'])->middleware(['auth:sanctum','access_control'])->name('".$routeNameDelete."');"));
         Artisan::call(command: 'make:migration create_' . Str::plural(strtolower($name)) . '_table --create=' . Str::plural(strtolower($name)));
     }
     protected function controller($name)
